+++ cli/tests/test_git_fetch.rs
@@ -1190,7 +1190,7 @@ fn test_git_fetch_removed_parent_branch() {
     "###);
 
     // Remove all branches in origin.
-    test_env.jj_cmd_ok(&source_git_repo_path, &["branch", "forget", "--glob", "*"]);
+    test_env.jj_cmd_ok(&source_git_repo_path, &["branch", "forget", "glob:*"]);
 
     // Fetch branches master, trunk1 and a1 from origin and check that only those
     // branches have been removed and that others were not rebased because of